Change tag format to be consistent with tagged type
