+++ wvlet-config/src/main/scala/wvlet/config/Config.scala
@@ -19,12 +19,10 @@ import java.util.Properties
 import wvlet.config.YamlReader.loadMapOf
 import wvlet.log.LogSupport
 import wvlet.log.io.IOUtil
-import wvlet.obj.ObjectBuilder.CanonicalNameFormatter
 import wvlet.obj._
 
 import scala.reflect.ClassTag
 import scala.reflect.runtime.{universe => ru}
-import scala.util.{Failure, Success, Try}
 
 case class ConfigHolder(tpe: ObjectType, value: Any)
 
@@ -74,59 +72,7 @@ object Config extends LogSupport {
       result
     }
   }
-
-  private[config] def extractPrefix(t: ObjectType): ConfigPrefix = {
-    def canonicalize(s: String): String = {
-      val name = s.replaceAll("Config$", "")
-      CanonicalNameFormatter.format(name)
-    }
-
-    t match {
-      case TaggedObjectType(raw, base, taggedType) =>
-        ConfigPrefix(Some(CanonicalNameFormatter.format(taggedType.name)), canonicalize(base.name))
-      case _ =>
-        ConfigPrefix(None, canonicalize(t.name))
-    }
-  }
-
-  case class ConfigPrefix(tag: Option[String], prefix: String)
-  case class ConfigParamKey(prefix: ConfigPrefix, param: String)
-  case class ConfigParam(key: ConfigParamKey, v: Any)
-
-  private[config] def configToProps(configHolder: ConfigHolder): Seq[ConfigParam] = {
-    val prefix = extractPrefix(configHolder.tpe)
-    val schema = ObjectSchema.of(configHolder.tpe)
-    val b = Seq.newBuilder[ConfigParam]
-    for (p <- schema.parameters) yield {
-      val key = ConfigParamKey(prefix, CanonicalNameFormatter.format(p.name))
-      Try(p.get(configHolder.value)) match {
-        case Success(v) => b += ConfigParam(key, v)
-        case Failure(e) =>
-          warn(s"Failed to read parameter ${p} from ${configHolder}")
-      }
-    }
-    b.result()
-  }
-
-  private[config] def toConfigKey(propKey: String): ConfigParamKey = {
-    val c = propKey.split("\\.")
-    c.length match {
-      case l if l >= 3 =>
-        val tag = c(0)
-        val prefix = c(1)
-        val param = CanonicalNameFormatter.format(c.drop(2).mkString)
-        ConfigParamKey(ConfigPrefix(Some(tag), prefix), param)
-      case l if l == 2 =>
-        val prefix = c(0)
-        val param = CanonicalNameFormatter.format(c(1))
-        ConfigParamKey(ConfigPrefix(None, prefix), param)
-      case other =>
-        throw new IllegalArgumentException(s"${propKey} should have ([tag].)?[prefix].[param] format")
-    }
-  }
 }
-
-import wvlet.config.Config._
 case class Config(env: Environment, configPaths: Seq[String], holder: Vector[ConfigHolder]) extends Iterable[ConfigHolder] with LogSupport {
   def getAll: Seq[ConfigHolder] = holder
   override def iterator: Iterator[ConfigHolder] = holder.iterator
@@ -154,7 +100,7 @@ case class Config(env: Environment, configPaths: Seq[String], holder: Vector[Con
     this + ConfigHolder(tpe, config)
   }
 
-  def registerFromYaml[ConfigType: ru.TypeTag : ClassTag](yamlFile: String) : Config = {
+  def registerFromYaml[ConfigType: ru.TypeTag : ClassTag](yamlFile: String): Config = {
     val tpe = ObjectType.ofTypeTag(implicitly[ru.TypeTag[ConfigType]])
     registerFromYaml(yamlFile, TypeUtil.newInstanceOf[ConfigType])
   }
@@ -186,28 +132,7 @@ case class Config(env: Environment, configPaths: Seq[String], holder: Vector[Con
   }
 
   def overrideWithProperties(props: Properties): Config = {
-    val overrides = {
-      import scala.collection.JavaConversions._
-      val b = Seq.newBuilder[ConfigParam]
-      for ((k, v) <- props) yield {
-        val key = toConfigKey(k)
-        b += ConfigParam(key, v)
-      }
-      b.result
-    }
-
-    val newConfigs = for (c <- holder) yield {
-      val configBuilder = ObjectBuilder.fromObject(c.value)
-      val prefix = extractPrefix(c.tpe)
-      val overrideParams = overrides.filter(_.key.prefix == prefix)
-      for (p <- overrideParams) {
-        trace(s"override: ${p}")
-        configBuilder.set(p.key.param, p.v)
-      }
-      ConfigHolder(c.tpe, configBuilder.build)
-    }
-
-    Config(env, configPaths, newConfigs)
+    ConfigOverwriter.overrideWithProperties(this, props)
   }
 
   def overrideWithPropertiesFile(propertiesFile: String): Config = {
